Instructions and explanation
TO USE THIS CODE, WHEN YOU ARE ON THE MAIN MENU, MAKE SURE THE CURSOR HIGHLIGHTS LOAD GAME BEFORE STARTING A NEW GAME. THIS CODE MUST BE ENABLED BEFORE STARTING THE GAME ON THE EMULATOR, AND IT WILL NOT BE FULLY DISABLED UNLESS YOU CLOSE THE GAME AFTER DISABLING THE CODE.

This is an AR code that does the following:
--Replaces the starting deck with a capture card that doesn't get removed from your deck or inventory when it captures something
--Disables all chest except for the Lucky Lion chest in Bridge of Sarvan
--Unlocks all levels
--Disables the Stone Golem obtained from Sol in Upper Chambers, the Mind Flayer in Hozo Town, and the trash quests in Kadishu Town
--Sets the starting gold amount to 1
--Stops capture cards from being used up when they capture an enemy 
--Buying capture cards is disabled


The idea of this challenge is to capture enough cards that you can beat the God of Harmony and/or Emperor with what you have captured. The card shops and Fairy House are optional.

Whenever you want to start a new run of this challenge mode, remember to make the cursor hover over Load Game before selecting New Game.